build(docker): Introduce the repo-cmd.py script and provide core Ubuntu Docker Images
This commit commit:

* Updates the gitignore file with the list of possible python
artifacts that we should ignore
* Adds python related pre-commit hookds
* Adds in the Makefile helper targets to create dev-containers based
on the infrastructure added in this commit.
* Adds the fodler ci-docker-images that contains the infrastructure
to build Ubuntu-22.04 docker images with and without GPU support
* Adds thes the repo-cmds.py script that calls docker-wrapper and
provides a CLI interface to interract with Docker and simplify
common Docker actions
* Updates the README.md with instructions how to use the above tools
and flows.
* Adds the environment-cfg.yml file that contains options passed
to repo-cmds.py and specifies configuration parameters specific
to the current compute environment.