build(llama.cpp): Add support for building the llama.cpp as an external CMake project
This commit:

* Checks out the llama.cpp repo by adding it to the west.yml file
* Renames the "example" chroot environment to llama-cpp
* Addresses build-dependency issues, where it splits the targets
that create the skeleton of the pyvenv from the original CMakeLists.txt
file under workspaces. This is necessary so that we can have targets
like llama-cpp that can install their artifacts inside the pyvenv.
For this reason the build-target and directory of the pyvenv need
to be known by the target that builds llama.cpp.
* Add helper Makefile command that lists all ninja targets of interest
* Update the documentation where necessary.